requirements.txt
opencv-python
ffmpeg-python
ffmpeg
numpy
soundfile
librosa
ultralytics
deepface
whisper
torch
transformers
openai
python-dotenv
accelerate
pyaudio

#museTalk
diffusers
torchvision
torchaudio
tensorflow
tensorboard
gdown
requests
imageio[ffmpeg]
omegaconf
gradio
spaces
moviepy

resampy
edge_tts
face_alignment
tqdm

#ernerf
torch-ema
ninja
trimesh
tensorboardX
pandas
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn
python_speech_features
numba
einops
configargparse
lpips

flask
flask_sockets
aiohttp
aiortc
aiohttp_cors